+++ src/shared/ExtensionMessage.ts
@@ -51,7 +51,9 @@ export interface ExtensionState {
 	uriScheme?: string
 	allowedCommands?: string[]
 	soundEnabled?: boolean
+	soundVolume?: number
 	diffEnabled?: boolean
+	debugDiffEnabled?: boolean
 }
 
 export interface ClineMessage {