+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Roo Cline Changelog
 
+## [2.2.13]
+
+-   Fixes to sound playing and applying diffs
+
+## [2.2.12]
+
+-   Better support for pure deletion and insertion diffs
+
+## [2.2.11]
+
+-   Added settings checkbox for verbose diff debugging
+
 ## [2.2.6 - 2.2.10]
 
 -   More fixes to search/replace diffs